add escrow contract

+++ src/Test20.sol
@@ -5,7 +5,7 @@ import "@openzeppelin/contracts/token/ERC20/ERC20.sol";
 contract Test20 is ERC20 {
     constructor() ERC20("Test", "TST") {}
 
-    function mint(uint amount) public {
+    function mint(uint256 amount) public {
         _mint(msg.sender, amount);
     }
 }
