+++ test/CrowdFund.t.sol
@@ -9,22 +9,24 @@ contract CrowdFundTest is Test {
     CrowdFund public crowdFund;
     Test20 public erc20Token;
 
-    uint INITIAL_ID = 1;
-    uint INTIAL_GOAL = 1000;
-    uint INITIAL_PLEDGE = 500;
+    uint256 INITIAL_ID = 1;
+    uint256 INTIAL_GOAL = 1000;
+    uint256 INITIAL_PLEDGE = 500;
 
     function setUp() public {
         crowdFund = new CrowdFund();
         erc20Token = new Test20();
-        crowdFund.launch(address(erc20Token), INTIAL_GOAL, uint32(block.timestamp + 1 hours), uint32(block.timestamp + 1 days));
+        crowdFund.launch(
+            address(erc20Token), INTIAL_GOAL, uint32(block.timestamp + 1 hours), uint32(block.timestamp + 1 days)
+        );
 
-        (address creator , , , , , ,) = crowdFund.campaigns(INITIAL_ID);
+        (address creator,,,,,,) = crowdFund.campaigns(INITIAL_ID);
         assertEq(creator, address(this));
     }
 
     function testCancel() public {
         crowdFund.cancel(INITIAL_ID);
-        (address creator , , , , , ,) = crowdFund.campaigns(INITIAL_ID);
+        (address creator,,,,,,) = crowdFund.campaigns(INITIAL_ID);
         assertEq(creator, address(0));
     }
 
@@ -35,7 +37,7 @@ contract CrowdFundTest is Test {
 
         erc20Token.approve(address(crowdFund), INITIAL_PLEDGE);
         crowdFund.pledge(INITIAL_ID, INITIAL_PLEDGE);
-        uint pledgeAmount = crowdFund.pledgedAmount(INITIAL_ID, address(321));
+        uint256 pledgeAmount = crowdFund.pledgedAmount(INITIAL_ID, address(321));
 
         assertEq(pledgeAmount, INITIAL_PLEDGE);
     }
@@ -49,8 +51,8 @@ contract CrowdFundTest is Test {
         crowdFund.pledge(INITIAL_ID, INITIAL_PLEDGE);
 
         crowdFund.unpledge(INITIAL_ID, INITIAL_PLEDGE);
-        uint tokenBalance = erc20Token.balanceOf(address(321));
-        uint pledgeAmount = crowdFund.pledgedAmount(INITIAL_ID, address(321));
+        uint256 tokenBalance = erc20Token.balanceOf(address(321));
+        uint256 pledgeAmount = crowdFund.pledgedAmount(INITIAL_ID, address(321));
 
         assertEq(pledgeAmount, 0);
         assertEq(tokenBalance, 500);
@@ -63,14 +65,14 @@ contract CrowdFundTest is Test {
         erc20Token.approve(address(crowdFund), INTIAL_GOAL);
         crowdFund.pledge(INITIAL_ID, INTIAL_GOAL);
 
-        uint pledgeAmount = crowdFund.pledgedAmount(INITIAL_ID, address(this));
+        uint256 pledgeAmount = crowdFund.pledgedAmount(INITIAL_ID, address(this));
         assertEq(pledgeAmount, INTIAL_GOAL);
 
         vm.warp(block.timestamp + 2 days);
         crowdFund.claim(INITIAL_ID);
 
-        (, , , , , , bool claimed) = crowdFund.campaigns(INITIAL_ID);
-        uint tokenBalance = erc20Token.balanceOf(address(this));
+        (,,,,,, bool claimed) = crowdFund.campaigns(INITIAL_ID);
+        uint256 tokenBalance = erc20Token.balanceOf(address(this));
 
         assertEq(tokenBalance, INTIAL_GOAL);
         assertEq(claimed, true);
@@ -85,11 +87,10 @@ contract CrowdFundTest is Test {
         crowdFund.pledge(INITIAL_ID, INITIAL_PLEDGE);
 
         crowdFund.refund(INITIAL_ID);
-        uint tokenBalance = erc20Token.balanceOf(address(321));
-        uint pledgeAmount = crowdFund.pledgedAmount(INITIAL_ID, address(321));
+        uint256 tokenBalance = erc20Token.balanceOf(address(321));
+        uint256 pledgeAmount = crowdFund.pledgedAmount(INITIAL_ID, address(321));
 
         assertEq(pledgeAmount, 0);
         assertEq(tokenBalance, 500);
     }
-
 }
